# docs/swagger-docs/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/users/profile": {
      "put": {
        "tags": [
          "User-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "operationId": "updateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "user",
                  "attributes": {
                    "email": "vishwasransing@gmail.com",
                    "mobile": "8805265888",
                    "password": "password",
                    "emailOtp": "123456",
                    "mobileOtp": "654321",
                    "ipAddress": "192.168.1.1",
                    "firstName": "Vishwas",
                    "lastName": "Ransingh",
                    "termsAccepted": "true",
                    "gender": "male",
                    "dob": "1998-05-05",
                    "salutation": "Mr."
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/signup": {
      "post": {
        "tags": [
          "User-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "user",
                  "attributes": {
                    "email": "vishwasransing@gmail.com",
                    "mobile": "8805265888",
                    "password": "password",
                    "emailOtp": "372725",
                    "mobileOtp": "372725",
                    "ipAddress": "192.168.1.1",
                    "firstName": "Vishwas",
                    "lastName": "Ransingh",
                    "termsAccepted": "true",
                    "gender": "male",
                    "dob": "1998-05-05",
                    "salutation": "Mr."
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/signIn": {
      "post": {
        "tags": [
          "User-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "user",
                  "attributes": {
                    "email": "vishwasransing@gmail.com",
                    "mobile": "8805265888",
                    "password": "password",
                    "emailOtp": "372725",
                    "mobileOtp": "372725",
                    "ipAddress": "192.168.1.1"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/password/reset": {
      "post": {
        "tags": [
          "User-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "operationId": "forgotPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "user",
                  "attributes": {
                    "mobile": "8805265888",
                    "email": "vishwasransing@gmail.com",
                    "mobileOtp": "584161",
                    "emailOtp": "584161",
                    "ipAddress": "192.168.1.1"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/otp/verify": {
      "post": {
        "tags": [
          "User-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserRequest"
            }
          }
        ],
        "operationId": "verifyOtpForUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestDTO"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "user",
                  "attributes": {
                    "mobile": "8805265888",
                    "email": "vishwasransing@gmail.com",
                    "mobileOtp": "584161",
                    "emailOtp": "584161"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/otp/generate-reset": {
      "post": {
        "tags": [
          "User-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "operationId": "generateOtpForPasswordResetForUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "user",
                  "attributes": {
                    "mobile": "7743898263",
                    "email": "vishwasransingh@gmail.com"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/otp/generate-register": {
      "post": {
        "tags": [
          "User-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "operationId": "generateOtpForNewRegistrationForUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "user",
                  "attributes": {
                    "mobile": "8805265888",
                    "email": "vishwasransing@gmail.com"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/otp/generate-login": {
      "post": {
        "tags": [
          "User-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "operationId": "generateOtpForLoginForUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "user",
                  "attributes": {
                    "mobile": "8805265888",
                    "email": "vishwasransing@gmail.com"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{id}": {
      "get": {
        "tags": [
          "User-Controller"
        ],
        "operationId": "retrieveUser",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User-Controller"
        ],
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {}
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/partners/banks/otp/generate": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "generateOtp",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "partnerId": "12345"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OTP generated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/partners/banks/otp/verify": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "verifyOtp",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "partnerId": "1",
                    "emailOtp": "123456",
                    "mobileOtp": "654321"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OTP verification successful",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/partners/banks/otp/generate-reset": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "generateOtpForPasswordReset",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "email": "partner@example.com",
                    "mobile": "9876543210"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OTP generated for password reset",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/partners/banks/otp/generate-register": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "generateOtpForNewRegistration",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "email": "partner@example.com",
                    "mobile": "9876543210"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OTP generated for new partner registration",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/partners/banks/otp/generate-login": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "generateOtpForLogin",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "email": "partner@example.com",
                    "mobile": "9876543210"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OTP generated for login",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RequestDTO": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RequestData"
          }
        }
      },
      "RequestData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "attributes": {
            "type": "object"
          }
        }
      },
      "DataDTOObject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "attributes": {
            "type": "object"
          }
        }
      },
      "ErrorDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "MetaDTOObject": {
        "type": "object",
        "properties": {
          "attributes": {
            "type": "object"
          }
        }
      },
      "ResponseDTOObject": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DataDTOObject"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaDTOObject"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDTO"
            }
          }
        }
      },
...
      "RequestId": {
        "type": "string",
        "pattern": "^[a-zA-Z0-9]",
        "description": "UUID for each request"
      },
      "UserRequest": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "attributes": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9+_.-]+@(.+)$",
                    "description": "User's email address"
                  },
                  "mobile": {
                    "type": "string",
                    "pattern": "^[1-9]\\d{9}$",
                    "description": "User's mobile number"
                  },
                  "password": {
                    "type": "string",
                    "pattern": "^[a-zA-Z0-9@$&]{5,50}$"
                  },
                  "emailOtp": {
                    "type": "string",
                    "pattern": "^[0-9]{6}$"
                  },
                  "mobileOtp": {
                    "type": "string",
                    "pattern": "^[0-9]{6}$"
                  },
                  "ipAddress": {
                    "type": "string"
                  },
                  "firstName": {
                    "type": "string",
                    "pattern": "^[a-zA-Z][a-z]{0,49}$"
                  },
                  "lastName": {
                    "type": "string",
                    "pattern": "^[a-zA-Z][a-z]{0,49}$"
                  },
                  "termsAccepted": {
                    "type": "string",
                    "pattern": "^(true|false)$"
                  },
                  "gender": {
                    "type": "string",
                    "pattern": "^(male|female|other)$"
                  },
                  "dob": {
                    "type": "string",
                    "format": "date"
                  },
                  "salutation": {
                    "type": "string",
                    "pattern": "^(Mr|Mrs)"
                  }
                }
              }
            }
          }
        },
        "required": [
          "data"
        ]
      },
...
    },
...
  }
}